[psxtools] arrayhex, [cvnds] map layout, default overlays, unused filelist

+++ git-update.sh
@@ -4,16 +4,6 @@ git=$(cat repo.git)
 ups=$(cat repo.fork)
 [ "$git" ] || exit
 
-msg="
-please type a comment to commit and push
-  type -retry  to retry git push [no commit]
-  type -pull   to do a git pull
-  type -update to update your forked repo
-  type -force  to overwrite the repo
-  type -last   to view the last 5 commits
-"
-[ $# = 0 ] && { echo "$msg"; exit; }
-
 git remote rm  origin
 git remote add origin "$git"
 if [ "$ups" ]; then
@@ -21,20 +11,53 @@ if [ "$ups" ]; then
 	git remote add upstream "$ups"
 fi
 
-case "$1" in
-	"-pull")
-		echo "git pull $git : master"
-		git pull origin master
+msg="
+usage: ${0##*/}  COMMAND  [files]...
+
+command
+  -push  COMMENT  add + commit + push updates to repo
+  -force CONFIRM  overwrite the repo
+  -pull    pull updates from repo
+  -push2   push updates to repo [SKIP add + commit]
+  -update  fetch updates from upstream
+  -last    view the last 5 commits
+
+if no command, list changes for upcoming update
+"
+##############################
+if [ $# = 0 ]; then
+	git diff HEAD
+	exit
+fi
+##############################
+cmd="$1"
+shift
+
+case "$cmd" in
+	'-push')
+		echo "git push $git : $@"
+		git add .
+		git ls-files --deleted -z | xargs -0 git rm
+		git reflog expire --expire=now --all
+		git gc --prune=now
+		git commit -m "$@"
+		git push origin master
 		;;
 	"-force")
 		[[ "$2" == "i_really_want_to_do_this" ]] || exit
 		echo "git push --force $git : master"
 		git push --force origin master
 		;;
-	"-retry")
+
+	"-pull")
+		echo "git pull $git : master"
+		git pull origin master
+		;;
+	"-push2")
 		echo "git push $git"
 		git push origin master
 		;;
+
 	"-update")
 		if [ "$ups" ]; then
 			echo "git fetch $ups"
@@ -47,12 +70,6 @@ case "$1" in
 		git log --pretty=oneline -5
 		;;
 	*)
-		echo "git push $git : $@"
-		git add .
-		git ls-files --deleted -z | xargs -0 git rm
-		git reflog expire --expire=now --all
-		git gc --prune=now
-		git commit -m "$@"
-		git push origin master
+		echo "$msg"
 		;;
 esac
